refactor(webpack.conf): move to config
